// src/shared/generated/RuntimeTabMoveResultRecord.ts
// This file was generated by [ts-rs](https://github.com/Aleph-Alpha/ts-rs). Do not edit this file manually.
import type { SystemRuntimeOperationSummaryRecord } from "./SystemRuntimeOperationSummaryRecord";

export type RuntimeTabMoveResultRecord = { targetWindowId: string, receipt: SystemRuntimeOperationSummaryRecord, };