// src/shared/generated/RuntimeWindowProvisionReceiptRecord.ts
// This file was generated by [ts-rs](https://github.com/Aleph-Alpha/ts-rs). Do not edit this file manually.
import type { EmbeddedLaunchTargetRecord } from "./EmbeddedLaunchTargetRecord";

export type RuntimeWindowProvisionReceiptRecord = { operationId: string, sourceWindowId: string, target: EmbeddedLaunchTargetRecord, windowGeneration: number, topologyRevision: number, };